+++ internal/search/run/tesutil_test.go
@@ -1,0 +1,11 @@
+package run
+
+import (
+	"github.com/sourcegraph/sourcegraph/cmd/frontend/backend"
+	"github.com/sourcegraph/sourcegraph/internal/database"
+)
+
+func resetMocks() {
+	database.Mocks = database.MockStores{}
+	backend.Mocks = backend.MockServices{}
+}
